core/src/main/java/org/opensbpm/engine/core/utils/LocalDateTimeAttributeConverter.java
/*
 * LocalDateTimeAttributeConverter.java
 *
 * Created on 01.01.2018,15:52:45
 *
 */
package org.opensbpm.engine.core.utils;

import java.sql.Timestamp;
import java.time.LocalDateTime;
import javax.persistence.AttributeConverter;
import javax.persistence.Converter;

/**
 * {@link AttributeConverter} to convert a {@link LocalDateTime} to a {@link Timestamp} and vice versa.
 */
@Converter(autoApply = true)
public class LocalDateTimeAttributeConverter implements AttributeConverter<LocalDateTime, Timestamp> {

    @Override
    public Timestamp convertToDatabaseColumn(LocalDateTime locDateTime) {
        return (locDateTime == null ? null : Timestamp.valueOf(locDateTime));
    }

    @Override
    public LocalDateTime convertToEntityAttribute(Timestamp sqlTimestamp) {
        return (sqlTimestamp == null ? null : sqlTimestamp.toLocalDateTime());
    }
}
